Separate files for model definitions and introduce model for user with password login
